test: split flaky tests on their own test suite
