+++ .github/workflows/ci.yaml
@@ -67,6 +67,28 @@ jobs:
         run: |
           make e2e-tests
 
+  e2e-tests-flaky:
+    name: E2E tests running
+    runs-on: ubuntu-latest
+    steps:
+      - name: Set up Go
+        uses: actions/setup-go@v3
+        with:
+          go-version: ${{ env.GOLANG_VERSION }}
+
+      - name: Set up Kind
+        uses: engineerd/setup-kind@v0.5.0
+        with:
+          version: "v0.23.0"
+          skipClusterCreation: true
+
+      - name: Checkout code
+        uses: actions/checkout@v3.1.0
+
+      - name: E2E tests running
+        run: |
+          make e2e-tests-flaky
+
   build:
     name: Build creation
     runs-on: ubuntu-latest
