+++ test/e2e/single_header_based_httproute_test.go
@@ -1,3 +1,5 @@
+//go:build flaky
+
 package e2e
 
 import (
