+++ test/e2e/single_tcproute_test.go
@@ -1,3 +1,5 @@
+//go:build !flaky
+
 package e2e
 
 import (
